pyproject.toml
[project]
name = "mlops-homework"
version = "0.1.0"
description = "Production-ready sentiment analysis FastAPI application for MLOps homework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clean-text>=0.7.1",
    "fastapi>=0.135.3",
    "joblib>=1.5.3",
    "pydantic-settings>=2.13.1",
    "scikit-learn==1.7.2",
    "sentence-transformers>=5.3.0",
    "torch>=2.11.0",
    "uvicorn>=0.43.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.20.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.9",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
